Cities/requirements.txt
streamlit
langchain
python-dotenv
duckduckgo-search
openai